infrastructure/dgtocean/bitcoin-cluster/docker/bitcoin/requirements.txt
kafka-python
python-dotenv
certifi
python-bitcoinrpc
pyzmq
requests
